Remove last traces of RobolectricPackageManager.
Use shadowOf(packageManager) instead.
